requirements.txt
django
mysql-connector-python
requests